requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.3.0
google-generativeai==0.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
flask-cors==4.0.0
